rested views
the changes I made earlier was to make it work with the new models. so it was mostly to test it.

will need to be changed a bit to work with new models, but should not be a lot of work.

tell me if I could help edit them later.